% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_api.R
\name{sa_api}
\alias{sa_api}
\title{Make a SellerActive API call}
\usage{
sa_api(endpoint = NULL)
}
\arguments{
\item{endpoint}{api endpoint}

\item{...}{additional named parameters}
}
\description{
Make a SellerActive API call
}
