% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_get_inventory.R
\name{sa_get_inventory}
\alias{sa_get_inventory}
\title{Get inventory from SellerActive}
\usage{
sa_get_inventory(sku = NULL)
}
\arguments{
\item{skus}{a single product SKU (character/scalar)}
}
\description{
Get inventory from SellerActive
}
