% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_parse_response.R
\name{sa_parse_response}
\alias{sa_parse_response}
\title{Parse a nested list returned by the API as a tibble}
\usage{
sa_parse_response(response)
}
\arguments{
\item{response}{a nested list}
}
\description{
Parse a nested list returned by the API as a tibble
}
